data-hub/requirements.txt
Flask>=3.1.2
ultralytics>=8.4.9
onnx==1.20.1
onnxruntime==1.23.2
opencv-python>=4.13.0.90
python-socketio[asyncio_client]